detect-tls-version: move unittests to tests/
